+++ examples/resources/identitycloud_cookie_domains/resource.tf
@@ -1,0 +1,3 @@
+resource "identitycloud_cookie_domains" "example" {
+  domains = ["mydomain1", "mydomain2"]
+}