+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 * **New Resource:** `identitycloud_variable` ([#7](https://github.com/pingidentity/terraform-provider-pingfederate/pull/7))
 * **New Resource:** `identitycloud_custom_domains` ([#8](https://github.com/pingidentity/terraform-provider-pingfederate/pull/8))
 * **New Resource:** `identitycloud_certificate` ([#9](https://github.com/pingidentity/terraform-provider-pingfederate/pull/9))
+* **New Resource:** `identitycloud_cookie_domains` ([#10](https://github.com/pingidentity/terraform-provider-pingfederate/pull/10))